requirements.txt
langchain
flask 
pypdf
python-dotenv
faiss-cpu
langchain-community
langchain_google_genai
langchain-experimental
-e . 